wip: Further work on the config override system.
