+++ com.playeveryware.eos/Runtime/Core/Common/WrappedStruct.cs
@@ -1,0 +1,48 @@
+/*
+ * Copyright (c) 2021 PlayEveryWare
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
+ */
+
+namespace PlayEveryWare.Common
+{
+    using System;
+
+    public abstract class WrappedStruct<T> : IEquatable<WrappedStruct<T>> where T : struct
+    {
+        protected T _value;
+
+        protected WrappedStruct(T value)
+        {
+            _value = value;
+        }
+
+        public override bool Equals(object obj)
+        {
+            return obj is WrappedStruct<T> wrappedStruct && Equals(wrappedStruct);
+        }
+
+        public abstract bool Equals(WrappedStruct<T> other);
+
+        public T Unwrap()
+        {
+            return _value;
+        }
+    }
+}
