+++ com.playeveryware.eos/Runtime/Core/Config/ProductConfig.cs
@@ -22,193 +22,8 @@
 
 namespace PlayEveryWare.EpicOnlineServices
 {
-    using Epic.OnlineServices.Platform;
+    using Common;
     using System;
-    using System.Collections;
-    using System.Collections.Generic;
-    using UnityEngine;
-    using UnityEngine.UIElements;
-
-    /// <summary>
-    /// Used to associate a name with a particular type.
-    /// </summary>
-    public class Named<T> : IEquatable<Named<T>>, IComparable<Named<T>>, IEquatable<T> where T : IEquatable<T>
-    {
-        /// <summary>
-        /// The name of the value (typically used for things like UI labels)
-        /// </summary>
-        public string Name;
-
-        /// <summary>
-        /// The value itself.
-        /// </summary>
-        public T Value;
-
-        public static Named<T> FromValue(T value, string name)
-        {
-            return new Named<T>() { Name = name, Value = value };
-        }
-
-        public int CompareTo(Named<T> other)
-        {
-            return string.Compare(Name, other.Name, StringComparison.Ordinal);
-        }
-
-        public bool Equals(T other)
-        {
-            if (Value == null && other == null)
-                return true;
-
-            if (Value == null)
-                return false;
-
-            return Value.Equals(other);
-        }
-
-        public override bool Equals(object obj)
-        {
-            return obj is Named<T> other && Equals(other);
-        }
-
-        public bool Equals(Named<T> other)
-        {
-            if (other is null)
-            {
-                return false;
-            }
-
-            if (ReferenceEquals(this, other))
-            {
-                return true;
-            }
-
-            return EqualityComparer<T>.Default.Equals(Value, other.Value);
-        }
-
-        public override int GetHashCode()
-        {
-            return HashCode.Combine(Name, Value);
-        }
-
-        public override string ToString()
-        {
-            return $"{Name} : {Value}";
-        }
-    }
-
-    public class Deployment : IEquatable<Deployment>
-    {
-        public readonly string SandboxId;
-
-        public Guid DeploymentId;
-
-        public bool Equals(Deployment other)
-        {
-            if (ReferenceEquals(null, other))
-            {
-                return false;
-            }
-
-            if (ReferenceEquals(this, other))
-            {
-                return true;
-            }
-
-            return SandboxId == other.SandboxId && DeploymentId.Equals(other.DeploymentId);
-        }
-
-        public override bool Equals(object obj)
-        {
-            if (ReferenceEquals(null, obj))
-            {
-                return false;
-            }
-
-            if (ReferenceEquals(this, obj))
-            {
-                return true;
-            }
-
-            if (obj.GetType() != this.GetType())
-            {
-                return false;
-            }
-
-            return Equals((Deployment)obj);
-        }
-
-        public override int GetHashCode()
-        {
-            return HashCode.Combine(SandboxId, DeploymentId);
-        }
-    }
-
-    public class SortedSetOfNamed<T> : SortedSet<Named<T>> where T : IEquatable<T>
-    {
-        public bool Add(string name, T value)
-        {
-            return !ContainsName(name) && base.Add(Named<T>.FromValue(value, name));
-        }
-
-        public bool ContainsName(string name)
-        {
-            foreach (Named<T> item in this)
-            {
-                if (item.Name == name)
-                {
-                    return true;
-                }
-            }
-
-            return false;
-        }
-
-        public bool Contains(T item)
-        {
-            Named<T> temp = Named<T>.FromValue(item, "temp");
-            return Contains(temp);
-        }
-    }
-
-    public abstract class WrappedStruct<T> : IEquatable<WrappedStruct<T>> where T : struct
-    {
-        protected T _value;
-
-        protected WrappedStruct(T value)
-        {
-            _value = value;
-        }
-
-        public override bool Equals(object obj)
-        {
-            return obj is WrappedStruct<T> wrappedStruct && Equals(wrappedStruct);
-        }
-
-        public abstract bool Equals(WrappedStruct<T> other);
-
-        public T Unwrap()
-        {
-            return _value;
-        }
-    }
-
-    public class WrappedClientCredentials : WrappedStruct<ClientCredentials>
-    {
-        public WrappedClientCredentials(ClientCredentials credentials) : base(credentials) { }
-
-        public override bool Equals(WrappedStruct<ClientCredentials> other)
-        {
-            if (other == null)
-            {
-                return false;
-            }
-
-            var temp = other.Unwrap();
-
-            return (temp.ClientId == _value.ClientId && 
-                    temp.ClientSecret == _value.ClientSecret);
-        }
-    }
 
     /// <summary>
     /// Contains information about the product entered by the user from the Epic
@@ -234,11 +49,11 @@ namespace PlayEveryWare.EpicOnlineServices
         public SortedSetOfNamed<WrappedClientCredentials> Clients;
 
         /// <summary>
-        /// The set of Sandboxes as defined within the Epic Developer Portal. For
-        /// EOS to function, at least one of these must be set, and it must
+        /// The set of Sandboxes as defined within the Epic Developer Portal.
+        /// For EOS to function, at least one of these must be set, and it must
         /// match the deployment indicated by the platform config.
         /// </summary>
-        public SortedSetOfNamed<string> Sandboxes;
+        public SortedSetOfNamed<SandboxId> Sandboxes;
 
         /// <summary>
         /// Backing field member for the Deployments property.
@@ -261,8 +76,10 @@ namespace PlayEveryWare.EpicOnlineServices
         {
             if (!Sandboxes.Contains(deployment.SandboxId))
             {
-                Sandboxes.Add(deployment.SandboxId)
+                Sandboxes.Add(deployment.SandboxId);
             }
+
+            _deployments.Add(name, deployment);
         }
 
         static ProductConfig()
@@ -270,6 +87,11 @@ namespace PlayEveryWare.EpicOnlineServices
             RegisterFactory(() => new ProductConfig());
         }
 
-        protected ProductConfig() : base("eos_product_config.json") { }
+        protected ProductConfig() : base("eos_product_config.json")
+        {
+            Clients = new SortedSetOfNamed<WrappedClientCredentials>("Client");
+            Sandboxes = new SortedSetOfNamed<SandboxId>("Sandbox");
+            _deployments = new SortedSetOfNamed<Deployment>("Deployment");
+        }
     }
 }
